Refactor open-world layer management and improve type safety
- Replaced integer layer masks with a new `worldOpenLayerMask_t` type for better clarity and type safety in open-world functions.
- Updated functions to utilize the new layer mask definitions, enhancing readability and maintainability of the code.
- Simplified boolean return statements in various functions for cleaner code.
- Improved handling of sector loading and unloading by using defined masks instead of bitwise operations.

These changes enhance the robustness of the open-world feature set, making the codebase easier to understand and work with.

+++ src/navigation/nav_recast.cpp
@@ -946,7 +946,7 @@ extern "C" qboolean Nav_BakeSectorTileToPath( const char *bspPath, const char *n
 	}
 	fclose( f );
 
-	if ( !Nav_BSP_ExtractFromSectorBuffer( buf, (int)length, cellX, cellY, sectorSize ) ) {
+	if ( !Nav_BSP_ExtractFromSectorBuffer( buf, (size_t)length, cellX, cellY, sectorSize ) ) {
 		free( buf );
 		return qfalse;
 	}
